Responses to copilot feedback including a connection fix for out of order reconnects, a fix for a big in the all reduce due to the wrong data type and some changes to the disposal patterns

+++ TensorSharp.Runtime/RedisKvBlockTier.cs
@@ -91,8 +91,8 @@ namespace TensorSharp.Runtime
                 return false;
             }
 
-            payload = new byte[payloadLength];
-            Array.Copy(raw, HeaderSize, payload, 0, payloadLength);
+            payload = new byte[(int)payloadLength];
+            Array.Copy(raw, HeaderSize, payload, 0, (int)payloadLength);
             return true;
         }
